app/pyproject.toml
[project]
name = "user-management-cognito"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "boto3>=1.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.3.0",
]
